requirements.txt
click==7.1.2
flashtext==2.7
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.3
pdf2image==1.13.1
Pillow==7.1.2
pyocr==0.7.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
